fix lint issues in helpers
